Use u32 to represent nested record entries number

+++ components/model/src/append_entry.rs
@@ -10,5 +10,5 @@ pub struct AppendEntry {
     pub offset: u64,
 
     /// Quantity of nested records
-    pub len: u16,
+    pub len: u32,
 }
